Add NOTICE file with copyright for collective work
The primary purpose of having Marvin Humphrey claim copyright over the
collective work is to clarify that though Mulver is spun off from Apache
Lucy, it is not a product of the Apache Software Foundation.

As with the copyright notices that ASF projects put in the NOTICE file,
this copyright notice only extends to the collective work.  Individual
contributors continue to maintain copyright over their contributions.